% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ah2sp}
\alias{ah2sp}
\title{Convert from ahull to spatial polygon}
\usage{
ah2sp(x, increment = 360, rnd = 10, proj4string = CRS(as.character(NA)))
}
\arguments{
\item{proj4string}{}
}
\value{

}
\description{
Convert from ahull to spatial polygon
}
