% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allee1954_3-data.R
\docType{data}
\name{allee1954_3}
\alias{allee1954_3}
\title{Allee1954-3 data set}
\format{An 10x10 aggression matrix}
\source{
\href{https://datadryad.org/resource/doi:10.5061/dryad.f76f2}{Shizuka
  D, McDonald DB (2015) Data from: The network motif architecture of
  dominance hierarchies. Dryad Digital Repository}
}
\usage{
data(allee1954_3)
}
\description{
Data from ...
}
\examples{
data(allee1954_3)
ec_power <- dom_ec(allee1954_3)
}
\references{
Allee, W. C., and J. C. Dickinson Jr. "Dominance and
  subordination in the smooth dogfish Mustelus canis (Mitchill)."
  Physiological Zoology 27.4 (1954): 356-364.
  (\href{https://doi.org/10.1086/physzool.27.4.30152372}{The University of Chicago Press})
}
\keyword{datasets}
