% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Compute bootstrap confidence intervals and correct for bootstrap bias}
\usage{
bootstrap_ci(
  bs_values,
  raw_mean,
  conf_level = 0.95,
  correction = FALSE,
  recenter = FALSE
)
}
\arguments{
\item{bs_values}{A vector of bootstrap estimates to be corrected}

\item{raw_mean}{The "raw" mean of the measure to be corrected, i.e., not the
bootsrapped estimate of the mean.}

\item{conf_level}{The confidence level. Default is 0.95, i.e. a 95%
confidence interval will be computed.}

\item{correction}{Controls whether to perform bias correction (TRUE) or not
(FALSE).}

\item{recetner}{Controls whetehr to center the confidence interval on the
value of `raw_mean`. Default is FALSE. Only relevant if `correction` is
FALSE.}
}
\value{
named vector with corrected "mean", "high" (upper CI limit), "low"
  (lower CI limit)
}
\description{
Compute bootstrap confidence intervals and correct for bootstrap bias
}
