% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{ComputeTransitionMatrix}
\alias{ComputeTransitionMatrix}
\title{Compute the transition matrix.}
\usage{
ComputeTransitionMatrix(A, eps = 1e-06)
}
\arguments{
\item{A}{aggression matrix}

\item{eps}{the epsilon parameter}
}
\value{
The transition matrix
}
\description{
The transition matrix is computed from the aggression matrix. The element T[i,j] of
the transition matrix is the probability that individual i attacks individual j.
}
