% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{correct_bootstrap_bias}
\alias{correct_bootstrap_bias}
\title{Correct for bootstrap bias in mean and hi and low estimates}
\usage{
correct_bootstrap_bias(bs_values, raw_mean)
}
\arguments{
\item{bs_values}{A vector of bootstrap estimates to be corrected}

\item{raw_mean}{The "raw" mean of the measure to be corrected, i.e., not the
bootsrapped estimate of the mean.}
}
\value{
named vector with corrected "mean", "high" (upper CI limit), "low"
  (lower CI limit)
}
\description{
Correct for bootstrap bias in mean and hi and low estimates
}
