% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dom_categorize_strategy}
\alias{dom_categorize_strategy}
\title{Categorize group-level strategy}
\usage{
dom_categorize_strategy(data, blur_data)
}
\arguments{
\item{blur_data}{}
}
\value{
String containing the name of the strategy used in the group.
}
\description{
Categorize group-level strategy
}
