% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{dom_focus}
\alias{dom_focus}
\title{Compute focus from an aggression matrix}
\usage{
dom_focus(aggression_matrix, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{An aggression matrix.}

\item{epsilon}{Regularization term for computing eigenvalue centrality.}
}
\value{
the computed focus value.
}
\description{
Compute focus from an aggression matrix
}
