% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{dom_make_blur_data}
\alias{dom_make_blur_data}
\title{Compute bootstrap estimates of focus and position for the downward null model
with confidence intervals}
\usage{
dom_make_blur_data(aggression_matrix, blur_values = 0:10/10,
  replications = 100, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{An aggression matrix.}

\item{blur_values}{A vector of blur values to use in the downward null model.}

\item{replications}{The number of bootstrap replications.}

\item{epsilon}{The regularization parameter for computing eigenvalue centrality.}
}
\value{
Returns a data frame with columns `blur`, `focus`, `position`
}
\description{
Compute bootstrap estimates of focus and position for the downward null model
with confidence intervals
}
