% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_models.R
\name{dom_make_downward_null}
\alias{dom_make_downward_null}
\title{Make a downward null model}
\usage{
dom_make_downward_null(aggression_matrix, blur = 0, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{An aggression matrix}

\item{blur}{The probability that an agent randomly attacks another
agent. If blur > 0 is means that the agent deviates from the pure
downward null heuristic}

\item{epsilon}{Regularization term for computing eigenvalue centrality in the
model}
}
\value{
A new aggression matrix based on the downward null heuristic model.
}
\description{
Make a downward null model
}
