% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{dom_position}
\alias{dom_position}
\title{Compute position from an aggression matrix}
\usage{
dom_position(aggression_matrix, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{An aggression matrix.}

\item{epsilon}{Regularization term for computing eigenvalue centrality.}
}
\value{
The position, i.e. the relative rank difference where the aggression is most focused.
}
\description{
Compute position from an aggression matrix
}
