% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_ranks}
\alias{dom_ranks}
\title{Get ranks from a vector of power scores}
\usage{
dom_ranks(power)
}
\arguments{
\item{power}{A vector of power scores (lower power is higher rank) for each
agent}
}
\value{
A vector of the ranks of each agent
}
\description{
Get ranks from a vector of power scores
}
