% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_resample}
\alias{dom_resample}
\title{Resample an aggression matrix}
\usage{
dom_resample(aggression_matrix)
}
\arguments{
\item{aggression_matrix}{The aggression matrix to be resampled}
}
\value{
a new aggression matrix resampled over the probability distribution
  obtained from the original aggression matrix
}
\description{
Resample an aggression matrix
}
