% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dom_transition_matrix}
\alias{dom_transition_matrix}
\title{Compute the transition matrix.}
\usage{
dom_transition_matrix(aggression_matrix, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{aggression matrix (square matrix)}

\item{epsilon}{the epsilon parameter (0 < eps <= 1)}
}
\value{
The transition matrix
}
\description{
The transition matrix is computed from the aggression matrix.
The element T[i,j] of the transition matrix is the probability
that individual i attacks individual j.
}
\examples{
dom_transition_matrix(matrix(c(0, 2, 3, 0), nrow = 2, ncol = 2))
}
