% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Compute the transition matrix.}
\usage{
transition_matrix(aggression_matrix, eps = 1e-12)
}
\arguments{
\item{aggression_matrix}{aggression matrix (square matrix)}

\item{eps}{the epsilon parameter (0 < eps <= 1)}
}
\value{
The transition matrix
}
\description{
The transition matrix is computed from the aggression matrix.
The element T[i,j] of the transition matrix is the probability
that individual i attacks individual j.
}
\examples{
transition_matrix(matrix(c(0, 2, 3, 0), nrow = 2, ncol = 2))
}
