% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment.R
\name{containment}
\alias{containment}
\title{Compute containment measure from transition matrix.}
\usage{
containment(transition_matrix)
}
\arguments{
\item{transition_matrix}{A transition matrix}
}
\value{
containment
}
\description{
Containment is computed as the second largest absolute eigenvalue of the
transition matrix.
}
\examples{
trans_matrix <- array(c(0, 0.5, 0.2, 1, 0, 0.8, 0, 0.5, 0), dim = c(3, 3))
containment(trans_matrix)
}
