% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containment.R
\name{containment}
\alias{containment}
\title{Compute containment measure from transition matrix.}
\usage{
containment(trans.mat)
}
\arguments{
\item{trans.mat}{}
}
\value{
containment
}
\description{
Containment is computed as the second left hand eigenvalue of the transition
matrix.
}
