% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dom_categorize_strategy}
\alias{dom_categorize_strategy}
\title{Categorize group-level strategy}
\usage{
dom_categorize_strategy(data, blur_data, use_data_ci = FALSE)
}
\arguments{
\item{use_data_ci}{Controls whether confidence intercal on data is used.
Default is FALSE, meaning only the point estimate is used.}
}
\value{
String containing the name of the strategy used in the group.
}
\description{
Categorize group-level strategy
}
