% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{dom_make_blur_data}
\alias{dom_make_blur_data}
\title{Compute bootstrap estimates of focus and position for the downward null
model with confidence intervals}
\usage{
dom_make_blur_data(
  aggression_matrix,
  blur_values = 0:10/10,
  replications = 100,
  epsilon = 0.694,
  conf_level = 0.95
)
}
\arguments{
\item{aggression_matrix}{An aggression matrix.}

\item{blur_values}{A vector of blur values to use in the downward null
model.}

\item{replications}{The number of bootstrap replications.}

\item{epsilon}{The regularization parameter for computing eigenvalue
centrality.}

\item{conf_level}{The confidence level used to compute lower and upper
limits of the confidence intervals on focus and position. The default
value is 0.95.}
}
\value{
Returns a data frame with columns `blur`, `focus`, `position`
  and confidence intervals on focus (`focus_ci_lo`, `focus_ci_hi`) and
  position (`position_ci_lo`, `position_ci_hi`).
}
\description{
Compute bootstrap estimates of focus and position for the downward null
model with confidence intervals
}
