% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{dom_make_data}
\alias{dom_make_data}
\title{Compute bootstrap estimates of focus and position for aggression matrix}
\usage{
dom_make_data(aggression_matrix, replications = 100, epsilon = 0.694)
}
\arguments{
\item{aggression_matrix}{An aggression matrix.}

\item{replications}{The number of bootstrap replications.}

\item{epsilon}{The regularization parameter for computing eigenvalue
centrality.}
}
\value{

}
\description{
Compute bootstrap estimates of focus and position for aggression matrix
}
